requirement.txt
asgiref             3.11.1
Django              6.0.2
djangorestframework 3.16.1
pillow              12.1.1
pip                 25.2
python-decouple     3.8
sqlparse            0.5.5
tzdata              2025.3
google-generativeai          0.8.6